shell-ui: Move the mock for useAuth to the setupTest
